parent_module/pyproject.toml
[project]
name = "parent-module"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-dotenv>=1.0.0",
    "subtraction-module @ file:///${PROJECT_ROOT}/../child_library/dist/subtraction_module-0.1.0.tar.gz",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["controllers", "handlers"]
